@@ -1,13 +1,17 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:webview_flutter/webview_flutter.dart';
 
-import 'package:webtrit_phone/data/package_info.dart';
+import 'package:webtrit_phone/data/data.dart';
+import 'package:webtrit_phone/extensions/extensions.dart';
 import 'package:webtrit_phone/l10n/l10n.dart';
 import 'package:webtrit_phone/widgets/widgets.dart';
+
+import 'webview_progress_indicator.dart';
 
 class WebViewScaffold extends StatefulWidget {
   const WebViewScaffold({
@@ -26,24 +30,93 @@ class WebViewScaffold extends StatefulWidget {
 }
 
 class _WebViewScaffoldState extends State<WebViewScaffold> {
-  late final WebViewController _controller;
+  final _webViewController = WebViewController();
+  final _progressStreamController = StreamController<int>.broadcast();
 
-  final StreamController<int> _progressStreamController = StreamController<int>.broadcast();
-  Timer? _resetProgressTimer;
+  Color? _backgroundColorCache;
+  Uri? _effectiveInitialUrlCache;
+
+  Uri _composeEffectiveInitialUrl() {
+    if (!widget.addLocaleNameToQueryParameters) {
+      return widget.initialUri;
+    } else {
+      return widget.initialUri.replace(
+        queryParameters: {
+          ...widget.initialUri.queryParameters,
+          'localeName': context.l10n.localeName,
+        },
+      );
+    }
+  }
 
   @override
-  Widget build(BuildContext context) {
-    final themeData = Theme.of(context);
-
-    final initialUriQueryParameters = Map.of(widget.initialUri.queryParameters);
-    if (widget.addLocaleNameToQueryParameters) {
-      initialUriQueryParameters['localeName'] = context.l10n.localeName;
-    }
-    final initialUrl = widget.initialUri.replace(queryParameters: initialUriQueryParameters).toString();
+  void initState() {
+    super.initState();
 
     final userAgent = '${PackageInfo().appName}/${PackageInfo().version} '
         '(${Platform.operatingSystem}; ${Platform.operatingSystemVersion})';
 
+    () async {
+      if (!kIsWeb) {
+        await Future.wait([
+          _webViewController.setUserAgent(userAgent),
+          _webViewController.enableZoom(false),
+          _webViewController.setJavaScriptMode(JavaScriptMode.unrestricted),
+          _webViewController.setNavigationDelegate(
+            NavigationDelegate(
+              onProgress: (progress) {
+                _progressStreamController.add(progress);
+              },
+            ),
+          ),
+        ]);
+      }
+    }();
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+
+    final themeData = Theme.of(context);
+    final backgroundColor = themeData.colorScheme.background;
+    if (_backgroundColorCache != backgroundColor) {
+      _backgroundColorCache = backgroundColor;
+      _webViewController.setBackgroundColor(backgroundColor);
+    }
+
+    final effectiveInitialUrl = _composeEffectiveInitialUrl();
+    if (_effectiveInitialUrlCache != effectiveInitialUrl) {
+      _effectiveInitialUrlCache = effectiveInitialUrl;
+      _webViewController.loadRequest(effectiveInitialUrl);
+    }
+  }
+
+  @override
+  void didUpdateWidget(WebViewScaffold oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    if (widget.initialUri != oldWidget.initialUri ||
+        widget.addLocaleNameToQueryParameters != oldWidget.addLocaleNameToQueryParameters) {
+      final effectiveInitialUrl = _composeEffectiveInitialUrl();
+      _effectiveInitialUrlCache = effectiveInitialUrl;
+      _webViewController.loadRequest(effectiveInitialUrl);
+    }
+  }
+
+  @override
+  void dispose() {
+    () async {
+      await _webViewController.setNavigationDelegate(NavigationDelegate());
+      await _webViewController.loadBlank();
+      await _progressStreamController.close();
+    }();
+
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: widget.title,
@@ -54,50 +127,21 @@ class _WebViewScaffoldState extends State<WebViewScaffold> {
               Icons.refresh,
             ),
             onPressed: () {
-              _controller.reload();
+              _webViewController.reload();
             },
           ),
         ],
-        bottom: PreferredSize(
-          preferredSize: const Size.fromHeight(2.0),
-          child: StreamBuilder<int>(
-            stream: _progressStreamController.stream,
-            builder: (context, snapshot) {
-              final snapshotData = snapshot.data;
-              if (snapshotData != null && snapshotData != -1) {
-                return LinearProgressIndicator(
-                  value: snapshotData / 100,
-                  backgroundColor: themeData.colorScheme.surface,
-                  minHeight: 2.0,
-                );
-              } else {
-                return const SizedBox();
-              }
-            },
-          ),
-        ),
       ),
-      body: WebView(
-        onWebViewCreated: (controller) => _controller = controller,
-        initialUrl: initialUrl,
-        javascriptMode: JavascriptMode.unrestricted,
-        onPageStarted: (_) {
-          _resetProgressTimer?.cancel();
-          _resetProgressTimer = null;
-        },
-        onPageFinished: (_) {
-          _resetProgressTimer = Timer(kTabScrollDuration, () {
-            _progressStreamController.add(-1);
-          });
-        },
-        onProgress: (progress) {
-          _progressStreamController.add(progress);
-        },
-        userAgent: userAgent,
-        zoomEnabled: false,
-        initialMediaPlaybackPolicy: AutoMediaPlaybackPolicy.always_allow,
-        allowsInlineMediaPlayback: true,
-        backgroundColor: themeData.colorScheme.background,
+      body: Stack(
+        alignment: AlignmentDirectional.topCenter,
+        children: [
+          WebViewWidget(
+            controller: _webViewController,
+          ),
+          WebViewProgressIndicator(
+            stream: _progressStreamController.stream,
+          ),
+        ],
       ),
     );
   }
